Move toolchain related functions to their own file
